@@ -9,8 +9,8 @@ public partial class SmartWindowHandler
 #if WINDOWS || MACCATALYST
         CommandMapper.AppendToMapping(nameof(ISmartWindow.SetPosition), SetPosition);
 
-        Mapper.AppendToMapping(nameof(ISmartWindow.MdiX), MapMdiXProperty);
-        Mapper.AppendToMapping(nameof(ISmartWindow.MdiY), MapMdiYProperty);
+        //Mapper.AppendToMapping(nameof(ISmartWindow.MdiX), MapMdiXProperty);
+        //Mapper.AppendToMapping(nameof(ISmartWindow.MdiY), MapMdiYProperty);
 
 #endif
     }
